requirements.txt
cocotb==1.8.1
pyuvm==2.8.0
cocotb-coverage==1.1.0
PyMuPDF==1.23.8
pyyaml==6.0.1
python-constraint==1.4.0
